+++ app/services/result.rb
@@ -1,0 +1,9 @@
+Result = Struct.new(:success, :object, :error_message) {
+  def success?
+    success == true
+  end
+
+  def failure?
+    !success
+  end
+}